examples: clean up chat
